ATR-855: moved all logic related to the events collection to the events collector class for a better separation of responsibilities
